connect: Reduce some duplication
Factor out most of the connect logic into a reusable function.
