+++ src/index.js
@@ -1,9 +1,7 @@
-import Url from 'url-parse';
 import WebSocket from 'ws';
 
-import log from './log';
+import connectImpl from './connect';
 import connectSSH from './transport/ssh';
-import XAPI from './xapi';
 import TSHBackend from './backend/tsh';
 import WSBackend from './backend/ws';
 import spawnTSH from './transport/tsh';
@@ -32,49 +30,13 @@ function initBackend(opts) {
   }
 }
 
-
-/**
- * Connect to an XAPI endpoint.
- *
- * @example
- * const xapi = connect('ssh://host.example.com:22');
- *
- * @param {string} url - Connection specification.
- * @param {Object} [options] - Connect options.
- * @param {string} [options.host] -
- *     Hostname to connect to.
- * @param {string} [options.username] -
- *     Username to authenticate with if protocol requires authentication.
- * @param {string} [options.password] -
- *     Password to authenticate with if protocol requires authentication.
- * @param {string} [options.loglevel] -
- *     Set the internal log level.
- * @return {XAPI} - XAPI interface connected to the given URI.
- */
-export function connect(url, options) { // eslint-disable-line import/prefer-default-export
-  if (arguments.length === 1 && typeof url === 'object') {
-    /* eslint-disable no-param-reassign */
-    options = url;
-    url = '';
-    /* eslint-enable */
-  }
-
-  const parsedUrl = new Url(url.match(/^\w+:\/\//) ? url : `ssh://${url}`);
-
+export function connect(url, options) {
   const opts = Object.assign({
     host: '',
     password: '',
     protocol: 'ssh:',
     username: 'admin',
     loglevel: 'warn',
-  }, parsedUrl, options);
-
-  opts.host = opts.hostname;
-  delete opts.hostname;
-
-  log.setLevel(opts.loglevel);
-  log.info('connecting to', url);
-
-  const backend = initBackend(opts);
-  return new XAPI(backend);
+  }, options);
+  return connectImpl(url, opts, initBackend);
 }
